src/Detection/Pneumonia/kaggle-rsna/requirements.txt
black>=19.10b0
Cython>=0.29.14
imgaug>=0.2.6
kaggle>=1.5.6
matplotlib>=3.1.2
numpy>=1.14.5
opencv-python>=3.4.2.17
pandas>=0.25.3
Pillow>=5.2.0
pretrainedmodels==0.7.4
pydicom>=1.3.0
pytest>=5.3.2
scikit-image>=0.16.2
scikit-learn>=0.19.2
scipy>=1.1.0
sklearn>=0.0
tensorboard>=1.15.0
torchsummary>=1.5.1
torchvision>=0.2.1
tqdm==4.19.9 













